backend/requirements.txt
fastapi
uvicorn
python-chess
pydantic
python-dotenv